requirements.txt
openai>=1.0.0
Pillow>=10.0.0
# Google Gemini API (google-generativeai)
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
# For HEIC image support
pillow-heif>=0.12.0
pyheif>=0.6.0
tqdm>=4.0.0 